% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformers.R
\name{addAttrCol}
\alias{addAttrCol}
\title{use biomaRt to translate annotation}
\usage{
addAttrCol(
  obj = dds,
  attrvec = c("hgnc_symbol", "ensembl_transcript_id"),
  filtvec = c("ensembl_transcript_id")
)
}
\description{
this function wraps biomaRt to get attributes
from the rownames of DESeqDataSet, DESeqResults,
and DESeqTransform objects, and adds them to the listData attribute
}
