% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaders.R
\name{checkRowNaming}
\alias{checkRowNaming}
\title{Row naming checks for tximport returned object}
\usage{
checkRowNaming(txobj, filelist)
}
\arguments{
\item{txobj}{the txi.rsem object loaded by loadRSEMs(). No default}

\item{filelist}{the filelist passed into loadFileList() for checking rownames against. No default}
}
\value{
List of matrices corresponding to count, abundance, and lengths for all the count tables indicated in your input file list (same as return from tximport), with corrected row names if they did not pass the checks
}
\description{
This function will check row naming for the txi.rsem object loaded by loadRSEMs() function
}
\details{
The function checks the first column of the matrices in the list returned by loadRSEMs (or another tximport return object) to ensure consistency in row names among files and the various matrices (counts, abundance, lengths)
Additionally, if the row names don't check out properly, this will rename them accordingly, but you should manually check that things are correct/consistent among the various matrices (txi.rsem$counts, txi.rsem$lengths, txi.rsem$abundance) in the list
This function generally should not be called directly, use the optional parameter checkRowNaming=TRUE into loadRSEMs() to call indirectly
}
\examples{
txobj <- checkRowNaming(txi.rsem,fileList)
}
\seealso{
\code{\link[=loadRSEMs]{loadRSEMs()}} which is the function that calls this validation function
}
