% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformers.R
\name{ddsTfm}
\alias{ddsTfm}
\title{transform expression data}
\usage{
ddsTfm(data = dds, tfm = "rlog", blind = FALSE, ...)
}
\arguments{
\item{data}{DESeqDataSet object to transform, Default is dds}

\item{tfm}{the transformation to perform, must be one of 'rlog' (for \code{\link[DESeq2:rlog]{DESeq2::rlog()}}) or 'varStbl' (for \code{\link[DESeq2:varianceStabilizingTransformation]{DESeq2::varianceStabilizingTransformation()}}). Default is 'rlog'}

\item{...}{any option that can be passed to \code{\link[DESeq2:rlog]{DESeq2::rlog()}} or \code{\link[DESeq2:varianceStabilizingTransformation]{DESeq2::varianceStabilizingTransformation()}}}
}
\value{
the transformed DESeqDataSet object as class DESeqTransform
}
\description{
wrapper function for transformations implemented by DESeq2 on count data for downstream analyses like PCA
}
\details{
wrapper function for transformations implemented by DESeq2,
either \code{\link[DESeq2:rlog]{DESeq2::rlog()}} or \code{\link[DESeq2:varianceStabilizingTransformation]{DESeq2::varianceStabilizingTransformation()}},
with options defaulting to those of 8OGseq paper usage (namely, blind=FALSE)
Note that you should loop over your ddsObjList as shown in the third example.
}
\examples{
dds.tfm <- ddsTfm(dds,'rlog',blind=FALSE)
dds.tfm <- ddsTfm(myDds,'varStbl')
ddsTfmList=c();for (NAME in names(ddsObjList)){ddsTfmList[[NAME]]<-ddsTfm(ddsObjList[[NAME]])}
}
\seealso{
\code{\link[=splitAndBuildDEobjs]{splitAndBuildDEobjs()}} which should be run before this to build the DESeqDataSet

\code{\link[DESeq2:rlog]{DESeq2::rlog()}} which this function wraps

\code{\link[DESeq2:varianceStabilizingTransformation]{DESeq2::varianceStabilizingTransformation()}} which this function wraps
}
