% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzers.R
\name{filterTopNpadj}
\alias{filterTopNpadj}
\title{Filter top N most significant diff transcripts of DESeqResults object}
\usage{
filterTopNpadj(
  res = ddr,
  num = 0.1 * nrow(res[!is.na(res$padj) & res$padj < 0.1, ])
)
}
\arguments{
\item{res}{the DESeqResults object from which to subset the top entries. Defaults to ddr}

\item{num}{the number of top entries to subset from the results object. Defaults to 10\% of the transcripts that have padj<0.1}
}
\value{
the top N subsetted DESeqResults object
}
\description{
subset the top N entries in the results of your modRIPseq analysis
}
\details{
this function cuts your results to the top N entries after sorting
Run \code{\link[=sortRes]{sortRes()}} to sort on your desired criteria before running this function to subset
Note that you should loop over your ddrObjList and pass each list entry therein to this function.
}
\examples{
ddrTop10pctSig <- filterTopNpadj(ddr,0.1*nrow(ddr[!is.na(ddr$padj)&ddr$padj<0.1,]))
filterTopNpadj(myddr,0.05*nrow(myddr[!is.na(myddr$padj)&myddr$padj<0.1,]))
for (NAME in names(ddrObjList)){ddrObjList[[NAME]]<-filterTopNpadj(ddrObjList[[NAME]])}
}
\seealso{
\code{\link[=sortRes]{sortRes()}} which should be run before this function to arrange the results according to your desired criteria, i.e. padj or log2FoldChange

\code{\link[=runDEseqSet]{runDEseqSet()}} that generates the list of DESeqResults objects as shown in the examples

\link[DESeq2:DESeqResults]{DESeq2::DESeqResults} that is the S4 object class this function will operate on
}
