% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaders.R
\name{loadFactors}
\alias{loadFactors}
\title{Load factors table for count table files}
\usage{
loadFactors(filepath = "./data/factors.tbl")
}
\arguments{
\item{filepath}{Path to the file containing your table of factors. Defaults to example file "./data/factors.tbl"}
}
\value{
Tibble object containing the factors table corresponding to filenames
}
\description{
This function will load the factors table file for annotating samples, which will be used in other package functions to set up analysis of count data from gene expression
}
\details{
The input file to this function should be a tab-separated table with:
first row containing titles for each column
first column should be the filename (i.e. filename or folder/filename), coresponding to the files listed on each line of the inputfile.txt file fed into the loadFileList function.
second column should be the antibody/pulldown treatment factor (i.e. Yes vs No)
third column should be the exposure condition (i.e. CAC/clean air/no vs OAM/exposure/yes)
fourth column should be the exposure level (i.e. high vs low), if you have multiple levels of exposure in your data
}
\examples{
loadFactors('./data/factors.tbl')
loadFactors('mydata/myfactorsTable.txt')
}
\seealso{
\code{\link[=splitAndBuildDEobjs]{splitAndBuildDEobjs()}} which follows this function in the modRIPseq pipeline
}
