% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaders.R
\name{loadFileList}
\alias{loadFileList}
\title{Load a file containing list of count table files}
\usage{
loadFileList(
  filepath = system.file("extdata", "inputfilelistAll.txt", package = "modRIPseq")
)
}
\arguments{
\item{filepath}{Path to the file containing list of filenames for RSEM output files. Defaults to "data/inputfilelist.txt"}
}
\value{
Vector of filenames pointing to the files containing RSEM count data
}
\description{
This function will load into memory a list of file paths for loading the actual count table files
}
\details{
The file input to this function should be a simple list of file paths to the output files from RSEM,
organized with each file path name on a newline, without any header row in the file.
the file paths in each newline should correspond to the first column in the factors table file passed as a parameter to the loadFactors function
}
\examples{
loadFileList("./data/inputfilelist.txt")
loadFileList("./myData/myFileList.txt")
}
\seealso{
[\code{\link[=loadFactors]{loadFactors()}}] which is the function that loads factor tables with first column corresponding to the fileList returned by this function
}
