% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaders.R
\name{loadRSEMs}
\alias{loadRSEMs}
\title{Load RSEM output files}
\usage{
loadRSEMs(filelist = fileList, checkRowNaming = FALSE)
}
\arguments{
\item{filelist}{Vector of filenames with each element containing a string that indicates the path to an rsem output count table file. Defaults to fileList.}

\item{checkRowNaming}{(OPTIONAL) Boolean flag to check all row names within files and across files to ensure row name consistency in counts, lengths, and abundances. Defaults to FALSE.}
}
\value{
List of matrices corresponding to count, abundance, and lengths for all the count tables indicated in your input file list (same as return from tximport)
}
\description{
This function will load the output count files from RSEM
}
\details{
Before running this, you should first execute loadFileList() to read the list of filenames.
In your inputfilelist.txt, ensure that your input files are all of the same output type, - or gene-level summaries.
This function will check that the filenames are consistent to ensure you have consistent analyses
Additionally, this function can call a secondary function for row name checking to ensure that all individual components
(counts, abundances, and lengths) are consistently named correctly within each file and across files
}
\examples{
txi.rsem <- loadRSEMs(fileLlist)
loadRSEMs(myFileList,TRUE)
}
\seealso{
\code{\link[tximport:tximport]{tximport::tximport()}} which this function wraps

\code{\link[=loadFileList]{loadFileList()}} which is the precursor to this function in the modRIPseq pipeline
}
