% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizTools.R
\name{makePCA}
\alias{makePCA}
\title{Make enrichr functional enrichment bar plot}
\usage{
makePCA(ddstfm = ddsTfm, intgroup = NULL)
}
\arguments{
\item{ddstfm}{the DESeqTransform object to be visualized on PCA plot}

\item{intgroup}{the intgroup option to pass into \code{\link[DESeq2:plotPCA]{DESeq2::plotPCA()}}. if not defined, the function will set the intgroup option from the "contrast" column of the passed ddstfm object. Default is NULL}
}
\value{
ggplot object containing the PCA plot of passed DESeqTransform object
}
\description{
This function wraps the \code{\link[DESeq2:plotPCA]{DESeq2::plotPCA()}}
function to examine data spread of modRIPseq data
Note that for looping over ddsTfmList object, to display the plot you'll need to wrap this in a call to \code{\link[base:plot]{base::plot()}} as in the third example
}
\details{
#suppl figure 8
#'@title Make log2FC 8OG enrichment plot with cutoff
#'
#'

#suppl fig 9a
#'@title
#'
#'

#make option for low (suppl fig 3a) vs high (suppl fig 10b)
#'@title Make venn diagram comparing high vs low level overlap
#'
#'

#fig 3c
#'@title Make log2FC heatmap with inset of top N in both high vs low
#'
#'

#not in paper? but part of data QC in pre-paper workup
#'@title Make MA plot
#'
#'

More details...
}
\examples{
makePCA(ddsTfm)
makePCA(myDdsTfm,intgroup="myDesignCol")
for (NAME in names(ddsTfmList)){plot(makePCA(ddsTfmList[[NAME]]))}
p <- makePCA(); p + theme(plot.title = element_text(hjust = 0.5))
}
\seealso{
\code{\link[DESeq2:plotPCA]{DESeq2::plotPCA()}} which this function wraps

\code{\link[=ddsTfm]{ddsTfm()}} which should be used to generate a DESeqTransform object for PCA plotting
}
