% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaders.R
\name{reorderFactorsByFile}
\alias{reorderFactorsByFile}
\title{Reorders factor tibble to match file list}
\usage{
reorderFactorsByFile(
  factortibble = factorTbl,
  filecol = file,
  filelist = fileList
)
}
\arguments{
\item{factortibble}{The tibble of factors from loadFactors() that has annotation corresponding to RSEM output file names. Defaults to factorTbl}

\item{filecol}{The column header in the tibble of factors that corresponds to the file names. Should be acceptable quoted or unquoted as passed parameter. Defaults to file}

\item{filelist}{The filelist vector returned from loadFileList() for reordering the tibble of factors to ensure proper annotation. Defaults to fileList}
}
\value{
The tibble of factors from loadFactors() that has been rearranged to match order of the file list vector from loadFileList()
}
\description{
This function will ensure the correct order of factors in the tibble according to the inputfilelist to properly match for DESeq analysis
}
\details{
This function is only called generally from within the \code{\link[=splitAndBuildDEobjs]{splitAndBuildDEobjs()}} function
}
\examples{
factorTbl <- reorderFactorsByFile(factorTbl,file,fileList)
}
\seealso{
\code{\link[=loadFileList]{loadFileList()}}

\code{\link[=loadFactors]{loadFactors()}}
}
