% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzers.R
\name{runDEseqSet}
\alias{runDEseqSet}
\title{Run DESeq2 analysis in the modRIPseq pipeline}
\usage{
runDEseqSet(ddsobjlist = ddsObjList, ...)
}
\arguments{
\item{ddsobjlist}{The named list of DESeq DataSet objects that have been created already by one of the DESeqDataSetFrom*** functions. Defaults to ddsObjList}

\item{...}{optionally, can pass named keyword parameters such as BPPARAM or parallel=TRUE for using biocparallel. for parallel=TRUE, you need to have registered the params already in your environment}

\item{formulalist}{The named list of vectors containing full and reduced formulas. The length of this vector should be . Defaults to formulaList}
}
\value{
named list of DESeq results objects that contain the DESeq2 analyzed DESeqDataSet objects
}
\description{
This function will run the DESeq analysis of the DESeqDataSet objects prepared in the modRIPseq pipeline to provide the resulting analyzed data.
}
\details{
This function provides a wrapper to the DESeq() function that analyzes DESeqDataSet objects.
The test used in this wrapper is hard-coded as the likelihood-ratio test, comparing a reduced model and a full model
The formula list parameter requires a specific structure, wherein for each name, the corresponding value is a vector with the first element as the reduced model and the second element as the full model in mosaic notation, i.e. formulaList[\link{'IPc'}]<- c('~Replicate','~Replicate+AbTreatment').
}
\examples{
runDEseqSet(ddsObjList,formulaList)
runDEseqSet(ddsObjList,formulaList,parallel=TRUE)
runDEseqSet(myDDSobjList,myFormulaList)
}
\seealso{
\code{\link[=splitAndBuildDEobjs]{splitAndBuildDEobjs()}} which is the step before this function in the 8OGseq pipeline

\code{\link[DESeq2:DESeq]{DESeq2::DESeq()}} which this function wraps

\link{BiocParallel} which can be used to employ parallel processing with DESeq2, and therefore with this package

\code{\link[BiocParallel:register]{BiocParallel::register()}} which must be used before passing parallel=TRUE parameter to this function
}
