% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaders.R
\name{setDDSnamesDose}
\alias{setDDSnamesDose}
\title{Set names for DESeq object list}
\usage{
setDDSnamesDose(factortibble = factorTbl)
}
\arguments{
\item{factortibble}{The tibble of factors that will be used to set DESeqDataSet object names in}
}
\value{
The list of names which will be used for denoting individual formulaList and ddsObjList entries
}
\description{
This function will dynamically set names for each DESeqDataSet based on factor tibble
}
\details{
This function will dynamically set names for the DESeqDataSet objects according to the requirements of the input files.
}
\examples{
setDDSnamesDose(factorTbl)
setDDSnames
}
\seealso{
[\code{\link[=loadFactors]{loadFactors()}}] which is a prerequisite for this function

[\code{\link[=splitAndBuildDEobjs]{splitAndBuildDEobjs()}}] which calls this function
}
