% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformers.R
\name{setTfmContrast}
\alias{setTfmContrast}
\title{add contrast information to ddsTfmList}
\usage{
setTfmContrast(ddstfmlist = ddsTfmList, ddrobjlist = ddrObjList)
}
\arguments{
\item{ddstfmlist}{the ddsTfmList on which you'd like to set the contrast values}

\item{ddrobjlist}{the ddrObjList to use for finding contrast to assign to the ddstfmlist. Defaults to ddrObjList}
}
\value{
the ddsTfmList object with each list entry containing the DESeqTransform object with an added column for contrast
}
\description{
This function operates on a list of DESeqTransform object to add contrast information for intgroup option of \code{\link[DESeq2:plotPCA]{DESeq2::plotPCA()}}
}
\details{
This function will use the passed ddrObjList parameter to identify
the contrast used for the corresponding entry
in the passed ddsTfmList parameter. This contrast which will be set as an
additional list entry in the DESeqTransform object, and this matches the
column name of the variable used in the design of the DESeq analysis.
The primary purpose here is to generate a linked value for
the intgroup option of PCA plot
}
\examples{
ddsTfmList <- setTfmContrast(ddsTfmList,ddrObjList)
myDdsTfmlist <- setTfmContrast(myDdsTfmlist,myDdrObjList)
}
\seealso{
\code{\link[=ddsTfm]{ddsTfm()}} which should be used to build the list of DESeqTransform objects this function operates on

\code{\link[=runDEseqSet]{runDEseqSet()}} which should be used to build the list of DESeqResults objects this function uses to get contrast data
}
