% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzers.R
\name{sortRes}
\alias{sortRes}
\title{Sort the DESeqResults object by column of choice}
\usage{
sortRes(res = ddr, colSort = padj, ...)
}
\arguments{
\item{res}{the DESeqResults object you wish to sort. Defaults to ddr}

\item{colSort}{the column you would like to sort on, will be passed to \link[dplyr:arrange]{dplyr::arrange} and so may be used as descending, i.e. desc(colSort). Defaults to padj}

\item{...}{Additional arguments to pass into \link[dplyr:arrange]{dplyr::arrange} for sorting the df, i.e. desc(log2FoldChange). No default value}
}
\value{
DESeqResults object that has been sorted according to chosen columns
}
\description{
Arrange the DESeqResults object by values in the chosen column(s)
}
\details{
This function will operate on a single DESeqResults object to sort the results by the chosen column(s).
It is suggested to sort on padj, the adjusted p-value, with secondary sorting on the log2FoldChange from high to low.
Note that you should loop over your ddrObjList and pass each list entry therein to this function.
}
\examples{
ddrSorted <- sortRes(ddr,padj)
sortRes(myddr,padj,desc(log2FoldChange))
for (NAME in names(ddrObjList)){ddrObjList[[NAME]]<-sortRes(ddrObjList[[NAME]])}
}
\seealso{
\code{\link[dplyr:arrange]{dplyr::arrange()}} that this method wraps

[\code{\link[=runDEseqSet]{runDEseqSet()}}] that generates the list of DESeqResults objects as shown in the examples

[\link[DESeq2:DESeqResults]{DESeq2::DESeqResults}] that is the S4 object class this function will operate on
}
