% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaders.R
\name{splitAndBuildDEobjs}
\alias{splitAndBuildDEobjs}
\title{Build DESeq2 objects for subsequent differential expression  & oxidation analysis}
\usage{
splitAndBuildDEobjs(factortibble = factorTbl, filelist = fileList)
}
\arguments{
\item{factortibble}{The reordered tibble of factors. Defaults to factorTbl}

\item{filelist}{The vector of filenames for rsem output files, same as the vector returned from loadFileList()}
}
\value{
List of DESeqDataSet objects, named with "input","IPc","IPx"
}
\description{
This function creates DESeq2 dataset objects for differential expression & oxidation analysis from the various file loading functions in the modRIPseq pipeline.
}
\details{
This function will use the other objects (factorTbl,fileList) created in the modRIPseq pipeline to create a set of DESeq dataset objects for subsequent analysis of exposure-induced differential expression and oxidation of transcripts.
}
\examples{
splitAndBuildDEobjs(factorTbl,fileList)
splitAndBuildDEobjs(myFactorTbl,myFileList)
}
\seealso{
\code{\link[DESeq2:DESeqDataSet]{DESeq2::DESeqDataSetFromTximport()}} which this function wraps

\code{\link[=loadFileList]{loadFileList()}} and \code{\link[=loadFactors]{loadFactors()}} which are the functions earlier in the modRIPseq pipeline that prepare inputs for this function
}
