% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizTools.R
\name{makeVenn}
\alias{makeVenn}
\title{#write to have option for high (suppl fig 10c) or low (suppl fig 9b) plot
#'@title Make volcano plot for modRIPseq results
#'@description This function will create a volcano plot for the DESeq results for modRIPseq data
#'@details
#'@param
#'@export
#'@return
#'@examples
#'@seealso}
\usage{
makeVenn(
  ddrsiglist = ddrSigList,
  samples = c("Input_High", "8OG_control_High", "8OG_OAM_High")
)
}
\description{
#not in paper? but part of data QC in pre-paper workup
#'@title Make independent filtering plot
#'@description
#'@details
#'@param
#'@export
#'@return
#'@examples
#'@seealso
}
\details{
#suppl figure 4a - can we get
#'@title Make Venn diagram plot of up/down regulated/oxidized transcripts
#'@description
#'@details
Note that you cannot pass more than 5 comparisons in as written,
i.e. you can't compare all 6 groups that are there by default
}
